Protect reinforced crops from water buckets
